+++ _assets/components/GuideDisplay.ts
@@ -1,7 +1,7 @@
-import {html, LitElement, nothing} from 'lit'
-import {customElement, property, state} from 'lit/decorators.js'
-import {Guide} from "../plugins";
-import {guides, proEdition, settings} from "../globals.ts";
+import { html, LitElement, nothing } from 'lit'
+import { customElement, property, state } from 'lit/decorators.js'
+import { Guide } from '../plugins'
+import { guides, proEdition, settings } from '../globals.ts'
 
 @customElement('guide-display')
 export class GuideDisplay extends LitElement {
@@ -31,7 +31,13 @@ export class GuideDisplay extends LitElement {
    * List of guides being displayed, along with their related information.
    */
   @state()
-  private _selectedGuide?: Guide = undefined;
+  private _selectedGuide?: Guide = undefined
+
+  /**
+   * Determines if TL;DR controls should appear.
+   */
+  @state()
+  private _showTldr = false
 
   /**
    * =========================================================================
@@ -39,7 +45,7 @@ export class GuideDisplay extends LitElement {
    * =========================================================================
    */
   private _selectGuide(slug: string) {
-    this._selectedGuide = this._guides.find(guide => guide.slug === slug)
+    this._selectedGuide = this._guides.find((guide) => guide.slug === slug)
   }
 
   /**
@@ -50,26 +56,34 @@ export class GuideDisplay extends LitElement {
   connectedCallback() {
     super.connectedCallback()
 
+    // Collect slugs from all rendered guides.
     const guideElements: NodeListOf<HTMLElement> = this.querySelectorAll('.guide')
-    const guideSlugs: string[] = [];
-    [...guideElements].forEach((guideElement) => {
+    const guideSlugs: string[] = []
+    ;[...guideElements].forEach((guideElement) => {
       if (guideElement.dataset.guideSlug) {
-        guideSlugs.push(guideElement.dataset.guideSlug);
+        guideSlugs.push(guideElement.dataset.guideSlug)
       }
-    });
-
-    this._guides = guideSlugs.map((slug) => {
-      return guides.find((guide) => guide.slug === slug);
     })
 
+    // Map rendered guides with guides in JavaScript object.
+    this._guides = guideSlugs.map((slug) => {
+      return guides.find((guide) => guide.slug === slug)
+    })
+
+    // Select the first guide to update navigation.
     if (this._guides.length) {
       this._selectedGuide = this._guides[0]
     }
 
-    if(proEdition && settings.enableGuideJavascript) {
+    // Detect TL;DR classes.
+    const tldrElements = this.querySelectorAll('.tldr-hide, .tldr-show')
+    this._showTldr = tldrElements.length > 0
+
+    // Fire callback for guide.
+    if (proEdition && settings.enableGuideJavascript) {
       guideSlugs.forEach((slug) => {
         try {
-          window.guideCallback[slug]?.();
+          window.guideCallback[slug]?.()
         } catch (error) {
           window.Craft.cp.displayError(this.tMessages.guideJsCallbackError.replace('[slug]', slug))
         }
@@ -85,42 +99,46 @@ export class GuideDisplay extends LitElement {
     const nav = html`
       <nav>
         <h2>${this.tMessages.guides}</h2>
-        <ul>${
-        this._guides.map((guide) => {
-          return html`<li class="${ this._selectedGuide?.slug === guide.slug ? 'selected' : nothing }">
-            <button type="button" @click="${() => this._selectGuide(guide.slug)}">${guide.title}</button>
-          </li>`
-        })
-      }</ul></nav>
+        <ul>
+          ${this._guides.map((guide) => {
+            return html`<li class="${this._selectedGuide?.slug === guide.slug ? 'selected' : nothing}">
+              <button type="button" @click="${() => this._selectGuide(guide.slug)}">${guide.title}</button>
+            </li>`
+          })}
+        </ul>
+      </nav>
     `
 
     const options = []
-    let tldrFound = false;
 
-    if (tldrFound) {
+    if (this._showTldr) {
       options.push(html`
         <div>
-          <input switch type="checkbox" id="tldr" name="tldr">
+          <input switch type="checkbox" id="tldr" name="tldr" />
           <label for="tldr">TL;DR <span class="info">HI</span></label>
         </div>
       `)
     }
 
     return html`
-      ${this._guides?.length > 1 || options.length ? html`
-      <aside>
-        ${this._guides?.length > 1 ? nav : nothing}
-        ${options.length ? options.map((option) => {
-          return option
-        }) : nothing}
-      </aside>
-      ` : nothing }
+      ${this._guides?.length > 1 || options.length
+        ? html`
+            <aside>
+              ${this._guides?.length > 1 ? nav : nothing}
+              ${options.length
+                ? options.map((option) => {
+                    return option
+                  })
+                : nothing}
+            </aside>
+          `
+        : nothing}
     `
   }
 
   protected willUpdate(changedProperties: Map<string, any>) {
     if (changedProperties.has('_selectedGuide')) {
-      const guideElements:NodeListOf<HTMLElement> = this.querySelectorAll('.guide')
+      const guideElements: NodeListOf<HTMLElement> = this.querySelectorAll('.guide')
 
       guideElements.forEach((guide) => {
         if (guide.dataset.guideSlug === this._selectedGuide?.slug) {